+++ openquake/calculators/ucerf_classical.py
@@ -21,7 +21,7 @@ import operator
 import numpy as np
 from openquake.baselib import parallel, general
 from openquake.hazardlib.calc.hazard_curve import classical
-from openquake.commonlib import readinput
+from openquake.commonlib.source_model_factory import random_filtered_sources
 from openquake.calculators import base
 from openquake.calculators.classical import (
     ClassicalCalculator, classical_split_filter)
@@ -63,8 +63,7 @@ class UcerfClassicalCalculator(ClassicalCalculator):
             [src] = grp
             srcs = list(src)
             if sample:
-                srcs = readinput.random_filtered_sources(
-                    srcs, self.src_filter, 1)
+                srcs = random_filtered_sources(srcs, self.src_filter, 1)
             acc = parallel.Starmap.apply(
                 classical_split_filter,
                 (srcs, self.src_filter, gsims, param, monitor),
