+++ openquake/commonlib/tests/logictree_test.py
@@ -35,6 +35,7 @@ from openquake.hazardlib import geo
 from openquake.baselib.general import gettemp
 from openquake.hazardlib.gsim import registry
 from openquake.commonlib import logictree, readinput, tests
+from openquake.commonlib.source_model_factory import SourceModelFactory
 from openquake.hazardlib.tom import PoissonTOM
 from openquake.hazardlib.pmf import PMF
 from openquake.hazardlib.mfd import TruncatedGRMFD, EvenlyDiscretizedMFD
@@ -2325,7 +2326,7 @@ class LogicTreeSourceSpecificUncertaintyTest(unittest.TestCase):
 
         mon = performance.Monitor()
         mags = [5.7, 5.98, 6.26, 6.54, 6.82, 7.1]
-        for sm in readinput.get_source_models(oqparam, gmc_lt, ssc_lt, mon):
+        for sm in SourceModelFactory(oqparam, gmc_lt, ssc_lt, mon).get_models():
             for src in sm.src_groups[0]:
                 if src.source_id == 'a2':
                     self.assertEqual(src.mfd.max_mag, 6.5)
